Implemented the assigning of rightValue and leftValue to random numbers in the override init of Question

+++ Maths/Question.h
@@ -11,8 +11,8 @@
 @interface Question : NSObject
 
 // Question properties
-@property (nonatomic) NSInteger rightValue;
-@property (nonatomic) NSInteger leftValue;
+@property (nonatomic) int rightValue;
+@property (nonatomic) int leftValue;
 @property (strong, nonatomic) NSString *question;
 @property (strong, nonatomic) NSString *answer;
